+++ CHANGELOG.md
@@ -58,7 +58,8 @@ This is, e.g., the case if the data user cuts out columns of a TSV without check
 # Changelist
 
 * next
-  * **Minor**: Added `maxFileAccessAttempts` and `fileAccessRetryWaitTimeMS` to make waiting time for delayed file appearance in NFS configurable. 
+  * **Minor**: Added `maxFileAppearanceAttempts` and `fileAppearanceRetryWaitTimeMS` to make waiting time for delayed file appearance configurable.  
+  * # TODO: This is probably not file appearance in the files system, but rather because of concurrency.  
   * **Note**: Introduced `dkfz_minimal_x.y.z` tags that mark release commits of the `dkfz_minimal` container.
     * **Patch**: Added `dkfz_minimal:1.1.0` container that contains now an almost full installation of packages present on old DKFZ CentOS 7.9 cluster nodes. Only few packages were left out. Skipped packages are marked with `#` comments in [`containers/dkfz_minimal/packages.txt`](containers/dkfz_minimal/packages.txt).
     * **Patch**: Added `dkfz_minimal:1.0.1` which fixes a missing module definition path. Thus, also `biomodal` and `workflows/husar modules` are becoming available in the container (when the DKFZ software stack is mounted). The container just adds a layer on the 1.0.0 container.